Unroll a confusing abstraction
This helps the code be more readable by removing indirection and a
number of unnecessary conditionals
