+++ robots/accession/technical_metadata.rb
@@ -1,17 +1,22 @@
 # frozen_string_literal: true
 
-# Initialize technicalMetadata
-
 module Robots
   module DorRepo
     module Accession
+      # Creates the technicalMetadata datastream
+      class TechnicalMetadata < Robots::DorRepo::Base
+        def initialize
+          super('dor', 'accessionWF', 'technical-metadata')
+        end
 
-      class TechnicalMetadata < AbstractMetadata
-        def self.params
-          { :process_name => 'technical-metadata', :datastream => 'technicalMetadata', :force => true }
+        def perform(druid)
+          obj = Dor.find(druid)
+          builder = DatastreamBuilder.new(object: obj,
+                                          datastream: obj.technicalMetadata,
+                                          force: true)
+          builder.build
         end
       end
-
     end
   end
 end
